+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -32,8 +32,8 @@ use crate::{
     wallet::{
         account::{
             operations::transaction::{
-                high_level::{create_alias::AliasOutputOptions, minting::mint_native_token::MintTokenTransactionDto},
-                prepare_output::OutputOptions,
+                high_level::{create_alias::CreateAliasParams, minting::mint_native_token::MintTokenTransactionDto},
+                prepare_output::OutputParams,
                 TransactionOptions,
             },
             types::{AccountBalanceDto, AccountIdentifier, TransactionDto},
@@ -43,7 +43,7 @@ use crate::{
             account_method::AccountMethod, dtos::AccountDetailsDto, message::Message, response::Response,
             AddressWithUnspentOutputsDto,
         },
-        AddressWithAmount, IncreaseNativeTokenSupplyOptions, NativeTokenOptions, NftOptions, Result, Wallet,
+        MintNativeTokenParams, MintNftParams, Result, SendAmountParams, Wallet,
     },
 };
 
@@ -532,18 +532,15 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::CreateAliasOutput {
-                alias_output_options,
-                options,
-            } => {
+            AccountMethod::CreateAliasOutput { params, options } => {
                 convert_async_panics(|| async {
-                    let alias_output_options = alias_output_options
-                        .map(|options| AliasOutputOptions::try_from(&options))
+                    let params = params
+                        .map(|options| CreateAliasParams::try_from(&options))
                         .transpose()?;
 
                     let transaction = account
                         .create_alias_output(
-                            alias_output_options,
+                            params,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -674,21 +671,13 @@ impl WalletMessageHandler {
             AccountMethod::IncreaseNativeTokenSupply {
                 token_id,
                 mint_amount,
-                increase_native_token_supply_options,
                 options,
             } => {
                 convert_async_panics(|| async {
-                    let increase_native_token_supply_options = match increase_native_token_supply_options {
-                        Some(native_token_options) => {
-                            Some(IncreaseNativeTokenSupplyOptions::try_from(&native_token_options)?)
-                        }
-                        None => None,
-                    };
                     let transaction = account
                         .increase_native_token_supply(
                             TokenId::try_from(&token_id)?,
                             U256::try_from(&mint_amount).map_err(|_| Error::InvalidField("mint_amount"))?,
-                            increase_native_token_supply_options,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -698,14 +687,11 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::MintNativeToken {
-                native_token_options,
-                options,
-            } => {
+            AccountMethod::MintNativeToken { params, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
                         .mint_native_token(
-                            NativeTokenOptions::try_from(&native_token_options)?,
+                            MintNativeTokenParams::try_from(&params)?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -728,14 +714,14 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::MintNfts { nfts_options, options } => {
+            AccountMethod::MintNfts { params, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
                         .mint_nfts(
-                            nfts_options
+                            params
                                 .iter()
-                                .map(NftOptions::try_from)
-                                .collect::<Result<Vec<NftOptions>>>()?,
+                                .map(MintNftParams::try_from)
+                                .collect::<Result<Vec<MintNftParams>>>()?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -745,13 +731,13 @@ impl WalletMessageHandler {
             }
             AccountMethod::GetBalance => Ok(Response::Balance(AccountBalanceDto::from(&account.balance().await?))),
             AccountMethod::PrepareOutput {
-                options,
+                params: options,
                 transaction_options,
             } => {
                 convert_async_panics(|| async {
                     let output = account
                         .prepare_output(
-                            OutputOptions::try_from(&options)?,
+                            OutputParams::try_from(&options)?,
                             transaction_options
                                 .as_ref()
                                 .map(TransactionOptions::try_from_dto)
@@ -762,17 +748,14 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::PrepareSendAmount {
-                addresses_with_amount,
-                options,
-            } => {
+            AccountMethod::PrepareSendAmount { params, options } => {
                 convert_async_panics(|| async {
                     let data = account
                         .prepare_send_amount(
-                            addresses_with_amount
+                            params
                                 .iter()
-                                .map(AddressWithAmount::try_from)
-                                .collect::<Result<Vec<AddressWithAmount>>>()?,
+                                .map(SendAmountParams::try_from)
+                                .collect::<Result<Vec<SendAmountParams>>>()?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -812,17 +795,14 @@ impl WalletMessageHandler {
             AccountMethod::SyncAccount { options } => Ok(Response::Balance(AccountBalanceDto::from(
                 &account.sync(options).await?,
             ))),
-            AccountMethod::SendAmount {
-                addresses_with_amount,
-                options,
-            } => {
+            AccountMethod::SendAmount { params, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
                         .send_amount(
-                            addresses_with_amount
+                            params
                                 .iter()
-                                .map(AddressWithAmount::try_from)
-                                .collect::<Result<Vec<AddressWithAmount>>>()?,
+                                .map(SendAmountParams::try_from)
+                                .collect::<Result<Vec<SendAmountParams>>>()?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -830,14 +810,11 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::SendNativeTokens {
-                addresses_and_native_tokens,
-                options,
-            } => {
+            AccountMethod::SendNativeTokens { params, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
                         .send_native_tokens(
-                            addresses_and_native_tokens.clone(),
+                            params.clone(),
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -845,14 +822,11 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::SendNft {
-                addresses_and_nft_ids,
-                options,
-            } => {
+            AccountMethod::SendNft { params, options } => {
                 convert_async_panics(|| async {
                     let transaction = account
                         .send_nft(
-                            addresses_and_nft_ids.clone(),
+                            params.clone(),
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;